// src/api/models/JourneyDefinitionResponse.ts
export interface JourneyDefinitionResponse {
    /**
     * 
     * @type {string}
     * @memberof JourneyDefinitionResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof JourneyDefinitionResponse
     */
    journeyCode?: string;
    /**
     * 
     * @type {string}
     * @memberof JourneyDefinitionResponse
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof JourneyDefinitionResponse
     */
    version?: number;
    /**
     * 
     * @type {Array<State>}
     * @memberof JourneyDefinitionResponse
     */
    states?: Array<State>;
    /**
     * 
     * @type {Array<TransitionResponse>}
     * @memberof JourneyDefinitionResponse
     */
    transitions?: Array<TransitionResponse>;
    /**
     * 
     * @type {boolean}
     * @memberof JourneyDefinitionResponse
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof JourneyDefinitionResponse
     */
    createdAt?: Date;
}